+++ Tests/PowerSyncTests/CrudTests.swift
@@ -199,4 +199,43 @@ final class CrudTests: XCTestCase {
         let finalValidationBatch = try await database.getCrudBatch(limit: 100)
         XCTAssertNil(finalValidationBatch)
     }
+    
+    func testCrudTransactions() async throws {
+        func insertInTransaction(size: Int) async throws {
+            try await database.writeTransaction { tx in
+                for _ in 0 ..< size {
+                    try tx.execute(
+                        sql: "INSERT INTO users (id, name, email) VALUES (uuid(), null, null)",
+                        parameters: []
+                    )
+                }
+            }
+        }
+        
+        // Before inserting any data, the iterator should be empty.
+        for try await _ in database.getCrudTransactions() {
+            XCTFail("Unexpected transaction")
+        }
+        
+        try await insertInTransaction(size: 5)
+        try await insertInTransaction(size: 10)
+        try await insertInTransaction(size: 15)
+        
+        var batch = [CrudEntry]()
+        var lastTx: CrudTransaction? = nil
+        for try await tx in database.getCrudTransactions() {
+            batch.append(contentsOf: tx.crud)
+            lastTx = tx
+            
+            if (batch.count >= 10) {
+                break
+            }
+        }
+        
+        XCTAssertEqual(batch.count, 15)
+        try await lastTx!.complete()
+        
+        let finalTx = try await database.getNextCrudTransaction()
+        XCTAssertEqual(finalTx!.crud.count, 15)
+    }
 }